api-shared/pyproject.toml
[project]
name = "api-shared"
version = "0.0.1"
description = "Shared dependencies for both api and the worker services."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hatchet-sdk[otel]>=1.27",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "opentelemetry-distro[otlp]>=0.52b0",
    "opentelemetry-instrumentation-logging>=0.52b0",
    "pydantic-settings>=2.8.1",
]

[project.optional-dependencies]
logfire = [
    "logfire[httpx,system-metrics]>=3.12.0",
]
langfuse = [
    "langfuse>=3.2.4",
]

[dependency-groups]
test = [
    "pytest>=8.3.3",
]

[tool.uv]
default-groups = ["test"]

[build-system]
requires = ["uv_build>=0.9,<0.14"]
build-backend = "uv_build"
